Pulled the code for IntervalMarker #4
So similarly like what I did for ValueMarker.java, I pulled the required code into the IntervalMarker.java and one point to note that I have reduced down the parameters since Marker object can be referenced by 'this' .
And in IntermediateAbstractRenderer. java, now we are just calling marker.getConcreateMarker(), which decides at run-time which marker code to be executed. #4